requirements.txt
fastapi
uvicorn
psycopg2
sqlalchemy
requests
python-dotenv
